Debugging updates
Mail delivery mission should work (it does in simulation). Now uses the user interface.
