+++ scripts/nodeLocations.json
@@ -1,5 +1,5 @@
-{"post1": [0, 2],
- "post2": [0, 1], 
- "post3": [1, 1], 
- "post4": [1, 0], 
- "post5": [1, 2]}
+{"post1": [0, 8],
+ "post2": [0, 4], 
+ "post3": [4, 4], 
+ "post4": [4, 0], 
+ "post5": [4, 8]}
